+++ src/libs/agent-runtime/types/chat.ts
@@ -90,8 +90,9 @@ export interface ChatStreamPayload {
 }
 
 export interface ChatCompetitionOptions {
-  callback: ChatStreamCallbacks;
+  callback?: ChatStreamCallbacks;
   headers?: Record<string, any>;
+  signal?: AbortSignal;
 }
 
 export interface ChatCompletionFunctions {
@@ -127,10 +128,3 @@ export interface ChatCompletionTool {
 }
 
 export type ChatStreamCallbacks = OpenAIStreamCallbacks;
-
-export interface OllamaChatMessage extends OpenAIChatMessage {
-  /**
-   * @description images for ollama vision models (https://ollama.com/blog/vision-models)
-   */
-  images?: string[];
-}