+++ src/services/topic/type.ts
@@ -1,10 +1,11 @@
 /* eslint-disable typescript-sort-keys/interface */
+import { BatchTaskResult } from '@/types/service';
 import { ChatTopic } from '@/types/topic';
 
 export interface CreateTopicParams {
   favorite?: boolean;
   messages?: string[];
-  sessionId: string;
+  sessionId?: string | null;
   title: string;
 }
 
@@ -16,7 +17,7 @@ export interface QueryTopicParams {
 
 export interface ITopicService {
   createTopic(params: CreateTopicParams): Promise<string>;
-  batchCreateTopics(importTopics: ChatTopic[]): Promise<any>;
+  batchCreateTopics(importTopics: ChatTopic[]): Promise<BatchTaskResult>;
   cloneTopic(id: string, newTitle?: string): Promise<string>;
 
   getTopics(params: QueryTopicParams): Promise<ChatTopic[]>;